pyproject.toml
[tool.poetry]
name = "infra-cli"
version = "0.1.0"
description = "Infrastructure as code"
authors = ["Bas Wilson <friedcatps@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
infra = "infra_cli.main:command"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
black = "^23.3.0"
pyyaml = "^6.0"
load-dotenv = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
